% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altdocwebrdemo-package.R
\docType{package}
\name{altdocwebrdemo-package}
\alias{altdocwebrdemo}
\alias{altdocwebrdemo-package}
\title{altdocwebrdemo: webR Demo for Generating WASM R Package Binaries}
\description{
Package to test out generating Web Assembly binaries for R packages using r-wasm GitHub Actions and deploying an interactive cell within an altdoc Quarto documentation site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rd.thecoatlessprofessor.com/quarto-webr-in-altdoc}
  \item \url{https://github.com/coatless-r-n-d/quarto-webr-in-altdoc}
  \item Report bugs at \url{https://github.com/coatless-r-n-d/quarto-webr-in-altdoc/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

}
\keyword{internal}
