% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demorwasmbinary-package.R
\docType{package}
\name{demorwasmbinary-package}
\alias{demorwasmbinary}
\alias{demorwasmbinary-package}
\title{demorwasmbinary: webR Demo for Generating WASM R Package Binaries}
\description{
Package to test out generating Web Assembly binaries for R packages using r-wasm GitHub Actions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-tutorials/webr-github-action-wasm-binaries}
  \item \url{https://tutorials.thecoatlessprofessor.com/webr-github-action-wasm-binaries/}
  \item Report bugs at \url{https://github.com/coatless-tutorials/webr-github-action-wasm-binaries/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

}
\keyword{internal}
