% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_surrealism.R
\docType{data}
\name{residual_surrealism}
\alias{residual_surrealism}
\title{Residual (Sur)Realism}
\format{
A data frame with 5,395 rows and 7 columns:
}
\source{
Leonard A Stefanski
\url{https://www.tandfonline.com/doi/abs/10.1198/000313007X190079}
}
\usage{
residual_surrealism
}
\description{
A dataset containing residuals from a surreal model.
}
\examples{
data(residual_surrealism)

model = lm(V1 ~ ., residual_surrealism)

plot(fitted(model),resid(model))
abline(0, 0, col = "red")

}
\keyword{datasets}
